Refactor viewer layout and markdown handling into modules
